Move utilities from core.py to _util.py
