+++ .gitignore
@@ -115,3 +115,4 @@ docs/source/collections/*
 pip-wheel-metadata/*
 tests/test_collections/*.nc
 .DS_Store
+dask-worker-space/
