+++ intake_esm/core.py
@@ -1,16 +1,14 @@
 import copy
 import json
 import logging
-from urllib.parse import urlparse
 
 import dask
 import fsspec
 import intake
 import numpy as np
 import pandas as pd
-import requests
 
-from ._util import logger
+from ._util import _fetch_and_parse_file, _get_dask_client, logger
 from .merge_util import _aggregate, _create_asset_info_lookup, _to_nested_dict
 
 
@@ -399,9 +397,11 @@ class esm_datastore(intake.catalog.Catalog):
         else:
             use_format_column = True
 
+        # replace path column with mapper (dependent on filesystem type)
         mapper_dict = {
-            path: _path_to_mapper(path, self.storage_options) for path in self.df[path_column_name]
-        }  # replace path column with mapper (dependent on filesystem type)
+            path: fsspec.get_mapper(path, **self.storage_options)
+            for path in self.df[path_column_name]
+        }
 
         groupby_attrs = []
         variable_column_name = None
@@ -585,71 +585,3 @@ def _load_group_dataset(
     )
 
     return group_id, ds
-
-
-def _is_valid_url(url):
-    """ Check if path is URL or not
-
-    Parameters
-    ----------
-    url : str
-        path to check
-
-    Returns
-    -------
-    bool
-    """
-    try:
-        result = urlparse(url)
-        return result.scheme and result.netloc and result.path
-    except Exception:
-        return False
-
-
-def _fetch_and_parse_file(input_path):
-    """ Fetch and parse ESMCol file.
-
-    Parameters
-    ----------
-    input_path : str
-            ESMCol file to get and read
-
-    Returns
-    -------
-    data : dict
-    """
-
-    data = None
-
-    try:
-        if _is_valid_url(input_path):
-            logger.info(f'Loading ESMCol from URL: {input_path}')
-            resp = requests.get(input_path)
-            data = resp.json()
-        else:
-            with open(input_path) as f:
-                logger.info(f'Loading ESMCol from filesystem: {input_path}')
-                data = json.load(f)
-
-    except Exception as e:
-        raise e
-
-    return data
-
-
-def _path_to_mapper(path, storage_options):
-    """Convert path to mapper"""
-    return fsspec.get_mapper(path, **storage_options)
-
-
-def _get_dask_client():
-    # Detect local default cluster already running
-    # and use it for dataset group loading.
-    client = None
-    try:
-        from distributed.client import _get_global_client
-
-        client = _get_global_client()
-        return client
-    except ImportError:
-        return client
